training-category middleware added

+++ routes/web.php
@@ -497,15 +497,15 @@ Route::middleware('auth')->group(function () {
         });
 
         Route::prefix('training-category')->group(function () {
-            Route::get('index', [TrainingCategoryController::class, 'index'])->name('training-category.index');
-            Route::get('create', [TrainingCategoryController::class, 'create'])->name('training-category.create');
-            Route::post('store', [TrainingCategoryController::class, 'store'])->name('training-category.store');
-            Route::get('edit/{ctrlno}', [TrainingCategoryController::class, 'edit'])->name('training-category.edit');
-            Route::put('update/{ctrlno}', [TrainingCategoryController::class, 'update'])->name('training-category.update');
-            Route::delete('destroy/{ctrlno}', [TrainingCategoryController::class, 'destroy'])->name('training-category.destroy');
-            Route::get('recentlyDeleted', [TrainingCategoryController::class, 'recentlyDeleted'])->name('training-category.recentlyDeleted');
-            Route::post('recently-deleted/restore/{ctrlno}', [TrainingCategoryController::class, 'restore'])->name('training-category.restore');
-            Route::delete('recently-deleted/force-delete/{ctrlno}', [TrainingCategoryController::class, 'forceDelete'])->name('training-category.forceDelete');
+            Route::get('index', [TrainingCategoryController::class, 'index'])->name('training-category.index')->middleware('checkPermission:compentency_training_category_view');
+            Route::get('create', [TrainingCategoryController::class, 'create'])->name('training-category.create')->middleware('checkPermission:compentency_training_category_add');
+            Route::post('store', [TrainingCategoryController::class, 'store'])->name('training-category.store')->middleware('checkPermission:compentency_training_category_add');
+            Route::get('edit/{ctrlno}', [TrainingCategoryController::class, 'edit'])->name('training-category.edit')->middleware('checkPermission:compentency_training_category_edit');
+            Route::put('update/{ctrlno}', [TrainingCategoryController::class, 'update'])->name('training-category.update')->middleware('checkPermission:compentency_training_category_edit');
+            Route::delete('destroy/{ctrlno}', [TrainingCategoryController::class, 'destroy'])->name('training-category.destroy')->middleware('checkPermission:compentency_training_category_delete');
+            Route::get('recentlyDeleted', [TrainingCategoryController::class, 'recentlyDeleted'])->name('training-category.recentlyDeleted')->middleware('checkPermission:compentency_training_category_delete');
+            Route::post('recently-deleted/restore/{ctrlno}', [TrainingCategoryController::class, 'restore'])->name('training-category.restore')->middleware('checkPermission:compentency_training_category_delete');
+            Route::delete('recently-deleted/force-delete/{ctrlno}', [TrainingCategoryController::class, 'forceDelete'])->name('training-category.forceDelete')->middleware('checkPermission:compentency_training_category_delete');
         });
 
         Route::prefix('training-secretariat')->group(function () {
